refactor: Ajout dossiers Calculator_Tests, Core_Calculator, Functions, User_interface
